refactor(package): constant 패키지로 상수 클래스 이동
