+++ src/test/java/lotto/OutputViewTest.java
@@ -2,7 +2,7 @@ package lotto;
 
 import lotto.model.Winning;
 import lotto.model.lotto.Lotto;
-import lotto.model.lotto.constant.LottoNumberPrintFormat;
+import lotto.constants.LottoNumberPrintFormat;
 import lotto.model.lotto.LottoChecker;
 import lotto.model.lotto.LottoPublisher;
 import lotto.model.number_generator.NumberGenerator;
